% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf.R
\name{adf}
\alias{adf}
\title{Conduct `ADF` unit root test using `EViews` routines}
\usage{
adf(series, info = "sic", caption = NULL, format = kable_format(), ...)
}
\arguments{
\item{series}{A vector of names or wildcard expressions for series object(s) contained in a dataframe.}

\item{info}{Name of the information criterion. For example, `SIC`, `AIC`, `HQ`.}

\item{caption}{Table caption as in `kable`.}

\item{format}{Table format in `kable`.}

\item{...}{Other arguments supported by `EviewsR` `import_kable()` function.}
}
\value{
An EViews workfile
}
\description{
Use this function to conduct unit root test using `EViews` routines
}
\examples{
library(URooTab)
\dontrun{
set.seed(1234)
x=rnorm(100)
y=cumsum(x)
z=cumsum(y)
dataFrame=data.frame(x,y,z)
adf(series=dataFrame,format="latex",info="aic")
}
}
\seealso{
Other important functions: 
\code{\link{pp}()}
}
\concept{important functions}
\keyword{documentation}
