\name{as.numeric-methods}
 \docType{methods}
  \alias{as.numeric-methods}
  \alias{as.numeric,vdate-method}
  \title{Methods for Function as.numeric in Package `uroot'}

  \description{
    Coerces an object of class `vdate' to a `numeric'.
  }

  \section{Methods}{
    \describe{
      \item{x = "vdate".}{Coerces an object of class `vdate' to a `numeric'.}
    }
  }

  \seealso{\code{\link{vdate-class}} and \code{\link{Mdates}}.}

\keyword{methods}
