\name{as.vdate-methods}
 \docType{methods}
  \alias{as.vdate-methods}
  \alias{as.vdate,ANY-method}
  \alias{as.vdate,ts-method}
  \title{Methods for Function as.vdate in Package `uroot'}

  \description{
    Coerces an object of class `ts' to a `vdate'.
  }

  \section{Methods}{
    \describe{
      \item{x = "vdate".}{Coerces an object of class `ts' to a `vdate'.}
    }
  }

  \seealso{\code{\link{vdate-class}}, \code{\link{Mdates}}, and \code{\link{stepdate-methods}}.}

\keyword{methods}
