\name{bgt.data}
\docType{data}
\alias{bgt.data}

\title{BGT-data Sample Data Set}

\description{
UK macroeconomic data.
Sample data set employed in Burridge, Gjorstrup and Taylor (2004).
}

\usage{
bgt.data
}

\format{
A list containing time series objects.
}

\references{
Burridge, P., Gjorstrup, F. and Taylor, R. (2004) 
\emph{Robust Inference on Seasonal Unit Roots via a Bootstrap Applied to OECD Macroeconomic Series}.
Department of Economics, City University London. Working Paper 04/08.
URL: \url{https://ideas.repec.org/p/cty/dpaper/04-08.html}.
}

\keyword{datasets}
