\name{ch.data}
\docType{data}
\alias{ch.data}

\title{CH-data Sample Data Set}

\description{
U.S. post World War II macroeconomic time series. The data set is described and employed 
in Canova and Hansen (JBES, 1995, Section 5.1).
}

\usage{
ch.data
}

\format{
A list containing time series objects.
}

\source{
\url{http://www.ssc.wisc.edu/~bhansen/progs/jbes_95.html}
}

\keyword{datasets}
