\name{dataFranses}
 \docType{data}
  \alias{dataFranses}

  \title{Franses(1996) data bank}
  \description{
      This data set contains some macroeconomic time series of the Basque Country.
   }

   \format{A .csv file.}

   \usage{
      dataFranses()
   }

   \source{P.H. Franses homepage: \url{http://www.few.eur.nl/few/people/franses/}}

   \references{
       P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996)
   }
   \keyword{datasets}


