\name{dataIKERBIDE}
 \docType{data}
  \alias{dataIKERBIDE}

  \title{Basque Country data bank}
  \description{
      This data set contains some macroeconomic time series of the Basque Country.
  }

  \format{A .csv file.}

  \usage{
      dataIKERBIDE()
  }

  \source{IKERBIDE databank by Gobierno Vasco}
  \keyword{datasets}
