\name{dataINE}
 \docType{data}
  \alias{dataINE}

  \title{Spain data bank}
  \description{
      This data set contains some macroeconomic time series of Spain.
   }

  \format{A .csv file.}

  \usage{
     dataINE()
  }

  \source{Instituto Nacional de Estadística de España. }
  \keyword{datasets}
