\name{maybeLags-class}
 \docType{class}
  \alias{maybeLags-class}
  \title{Class "maybeLags" ~~~ }

  \description{A class for internal use to. It is the union of the classes "NULL", "list", "character",
    "numeric", and "matrix".}

  \section{Objects from the Class}{A virtual Class: No objects may be created from it.}

  \section{Methods}{
    No methods defined with class "maybeLags" in the signature.
  }

  \author{Javier López-de-Lacalle \email{javlacalle@yahoo.es} and Ignacio Díaz-Emparanza
    \email{Ignacio.Diaz-Emparanza@ehu.es}}

  \seealso{\code{\link{maybeRegvar-class}}.}

\keyword{classes}
