\name{maybeRegvar-class}
 \docType{class}
  \alias{maybeRegvar-class}
  \title{"maybeRegvar" Class}

  \description{A class for internal use to. It is the union of the classes "NULL", "numeric", and "matrix".}

  \section{Objects from the Class}{A virtual Class: No objects may be created from it.}

  \section{Methods}{
    No methods defined with class "maybeRegvar" in the signature.
  }

  \references{ ~put references to the literature/web site here ~ }

  \author{Javier López-de-Lacalle \email{javlacalle@yahoo.es} and Ignacio Díaz-Emparanza
    \email{Ignacio.Diaz-Emparanza@ehu.es}}

  \seealso{\code{\link{maybeLags-class}}.}

\keyword{classes}
