\name{mccapv}
 \docType{data}
  \alias{mccapv}

  \title{Macroeconomic Time Series from the CAPV}
  \description{
      This data set contains the macroeconomic time series from the Comunidad Autnoma del Pas Vasco [CAPV].
   }

   \format{An .rda file which stores a list with objects of class \code{ts}.}

   \usage{
      data("mccapv")
   }

   \details{
     Available time series: (Each corresponding label in the list object appears in quotes.)

     \itemize{
       \item "airp": Airpassengers. Sample: 1982.1-2003.10. Monthly.
       \item "epaact": Actives. Sample: 1977.1-2004.2. Quarterly.
       \item "iai": Industrial activity index. Sample: 1984.7-2003.6. Monthly.
       \item "ipc": Consumer price index. Sample: 1993.1-2001.12. Monthly.
       \item "ipi": Industrial production index. Sample: 1990.1-2003.11. Monthly.
       \item "ipri": Industrial price index. Sample: 1993.1-2003.12. Monthly.
       \item "licof": Official bidding. Sample: 1985.1-2003.3. Quarterly.
       \item "mtur": Private car registration. Sample: 1980.1-2003.12. Monthly.
       \item "mvic": Industrial cargo vehicles registration. Sample: 1980.1-2003.12. Monthly.
       \item "paroreg": Registered unemployment. Sample: 1980.1-2003.12. Monthly.
       \item "pcemen": Cement production. Sample: 1980.1-2001.12. Monthly.
       \item "pernhot": Hotel occupation. Sample: 1981.1-2003.12. Monthly.
       \item "valorpr": Production value. Sample: 1989.1-2003.3. Quarterly.
       \item "vpoinic": Initiated Council houses. Sample: 1982.1-2003.12. Monthly.
       \item "vpoterm": Finished Council houses. Sample: 1982.1-2003.12. Monthly.
     }
   }

   \source{P.H. Franses's homepage: \url{http://www.few.eur.nl/few/people/franses/}.}

   \references{
      P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).
   }
   \keyword{datagen}


