\name{show-methods}
 \docType{methods}
  \alias{show-methods}
  \alias{show,adfstat-method}
  \alias{show,chstat-method}
  \alias{show,chrecst-method}
  \alias{show,hegystat-method}
  \alias{show,kpssstat-method}
  \alias{show,kpssrecst-method}
  \alias{show,adfrecst-method}
  \alias{show,hegyrecst-method}
  \alias{show,vdate-method}
  \title{Methods for Function show in Package `uroot'}

  \description{
    Methods for Function show in Package `urootlast' \pkg{uroot}.
  }

  \section{Methods}{
    \describe{
      \item{object = "adfstat".}{Show the ADF test statistic.}
      \item{object = "adfrecst".}{Show the relevant information from the ADF recursive test.}
      \item{object = "chstat".}{Show the relevant information from the CH test.}
      \item{object = "chrecst".}{Show the relevant information from the CH recursive test.}
      \item{object = "hegystat".}{Show the the HEGY test statistics.}
      \item{object = "hegyrecst".}{Show the relevant information from the HEGY recursive test.}
      \item{object = "kpssstat".}{Show the relevant information from the KPSS test.}
      \item{object = "kpssrecst".}{Show the relevant information from the KPSS recursive test.}
      \item{object = "vdate".}{Show the date related to and observation and vice versa.}
    }
  }

  \seealso{\code{\link{ADF.test}}, \code{\link{ADF.rectest}}, \code{\link{KPSS.test}},
    \code{\link{KPSS.rectest}}, \code{\link{HEGY.test}}, \code{\link{HEGY.rectest}},
    \code{\link{CH.test}}, and \code{\link{CH.rectest}}.}

\keyword{methods}
