\name{summary-methods}
 \docType{methods}
  \alias{summary-methods}
  \alias{summary,adfstat-method}
  \alias{summary,hegystat-method}
  \title{Methods for Function summary in Package `uroot'}

  \description{
    Methods for Function summary in Package `urootlast' \pkg{uroot}.
  }

  \section{Methods}{
    \describe{
      \item{object = "adfstat".}{Summarize the ADF test regression.}
      \item{object = "hegystat".}{Summarize the the HEGY test regression.}
    }
  }

  \seealso{\code{\link{ADF.test}}, and \code{\link{HEGY.test}}.}

\keyword{methods}
