% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroot.R
\name{uroot}
\alias{uroot}
\title{Conduct unit root test using `EViews` routines}
\usage{
uroot(x, test = c("adf", "pp"), info = "sic")
}
\value{
An EViews workfile
}
\description{
Use this function to conduct unit root test using `EViews` routines
}
\examples{
library(EviewsR)
\dontrun{
demo(exec_commands)

eviews_graph(wf="exec_commands",page = "eviewspage1",series="x y",mode = "overwrite",
graph_options = "m")

# Create graph(s) from dataframe

Data=data.frame(x=cumsum(rnorm(100)),y=cumsum(rnorm(100)))

eviews_graph(series=Data,start_date=1990,frequency="m")

# Create graphs in one frame (group=TRUE)

eviews_graph(series=Data,group=TRUE,start_date="1990Q4",frequency="Q")
}
}
\concept{important functions}
\keyword{documentation}
