% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Conduct unit root test using `EViews` routines}
\usage{
pp(series, info = "sic", caption = NULL, format = kable_format(), ...)
}
\arguments{
\item{series}{A vector of names or wildcard expressions for series object(s) contained in a dataframe.}

\item{info}{Name of the information criterion. For example, `SIC`, `AIC`, `HQ`.}

\item{caption}{Table caption as in `kable`.}

\item{format}{Table format in `kable`.}

\item{...}{Other arguments supported by `EviewsR` `import_kable()` function.}
}
\value{
An EViews workfile
}
\description{
Use this function to conduct `PP` unit root test using `EViews` routines and report it in a table.
}
\examples{
library(URooTab)
\dontrun{
Data=data.frame(x=cumsum(rnorm(100)),y=cumsum(rnorm(100)))
pp(series=Data,test="ADF",info="sic")

}
}
\seealso{
Other important functions: 
\code{\link{adf}()}
}
\concept{important functions}
\keyword{documentation}
