% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroot.R
\name{uroot}
\alias{uroot}
\title{Conduct unit root test using `EViews` routines}
\usage{
uroot(
  series,
  test = c("adf", "pp"),
  info = "sic",
  caption = NULL,
  format = kable_format(),
  ...
)
}
\arguments{
\item{series}{A vector of names or wildcard expressions for series object(s) contained in a dataframe.}

\item{test}{Name of the unit root test. For example, `ADF`, `PP`.}

\item{info}{Name of the information criterion. For example, `SIC`, `AIC`, `HQ`.}

\item{caption}{Table caption as in `kable`.}

\item{format}{Table format in `kable`.}

\item{...}{Other arguments supported by `EviewsR` `import_kable()` function.}
}
\value{
An EViews workfile
}
\description{
Use this function to conduct unit root test using `EViews` routines
}
\examples{
library(UROOT)
\dontrun{
Data=data.frame(x=cumsum(rnorm(100)),y=cumsum(rnorm(100)))
uroot(series=Data,test="ADF",info="sic")

}
}
\keyword{documentation}
